refactor: MatchingStatus enum 의 위치 변경
